backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
google-generativeai
pydantic
python-dotenv
pyjwt
passlib[bcrypt]
